fix(api): move truncateSignatures to actuator layer
