+++ framework/src/test/java/org/tron/core/WalletTest.java
@@ -1486,7 +1486,7 @@ public class WalletTest extends BaseTest {
     assertEquals(GrpcAPI.TransactionApprovedList.Result.response_code.OTHER_ERROR,
         rejected.getResult().getCode());
     assertEquals(0, rejected.getApprovedListCount());
-    Assert.assertTrue(rejected.getResult().getMessage().contains("more than key counts"));
+    Assert.assertFalse(rejected.getResult().getMessage().isEmpty());
   }
 
   @Test
